+++ lib/src/chunking/pointDistribution.cu
@@ -7,9 +7,9 @@ __global__ void kernelDistributing(
         Chunk *grid,
         Vector3 *cloud,
         Vector3 *treeData,
-        uint64_t *tmpIndexRegister,
+        uint32_t *tmpIndexRegister,
         PointCloudMetadata metadata,
-        uint64_t gridSize
+        uint32_t gridSize
         ) {
 
     int index = blockIdx.x * blockDim.x + threadIdx.x;
@@ -20,7 +20,7 @@ __global__ void kernelDistributing(
 
     auto gridIndex = tools::calculateGridIndex(point, metadata, gridSize);
 
-    uint64_t dstIndex = gridIndex;
+    uint32_t dstIndex = gridIndex;
     bool isFinished = grid[dstIndex].isFinished;
 
     while(!isFinished) {
@@ -28,14 +28,14 @@ __global__ void kernelDistributing(
         isFinished = grid[dstIndex].isFinished;
     }
 
-    uint64_t i = atomicAdd(&tmpIndexRegister[dstIndex], 1);
+    uint32_t i = atomicAdd(&tmpIndexRegister[dstIndex], 1);
     treeData[grid[dstIndex].chunkDataIndex + i] = point;
 }
 
 void PointCloud::distributePoints() {
 
     // Create temporary indexRegister for assigning an index for each point within its chunk area
-    auto tmpIndexRegister = make_unique<CudaArray<uint64_t>>(itsCellAmount, "tmpIndexRegister");
+    auto tmpIndexRegister = make_unique<CudaArray<uint32_t>>(itsCellAmount, "tmpIndexRegister");
 
     // Calculate kernel dimensions
     dim3 grid, block;
